Tweak symlink to do the heavy lifting.

+++ setup-symlink.sh
@@ -1,26 +1,31 @@
+set -e
+
 # Variables
 dir=~/.dotfiles
-files=".zshrc"
-
-# Lets make it look nicer
-echo ""
-echo "- - - - - - - - - -"
-echo ""
+files=(
+  ".zshrc"
+  ".oh-my-zsh/custom/themes"
+  ".oh-my-zsh/custom/zsh-autosuggestions.zsh"
+)
 
 # change to the dotfiles directory
 echo "Changing to the $dir directory"
 cd $dir
 
-# Lets make it look nicer
-echo ""
-echo "- - - - - - - - - -"
-echo ""
-
 # Create symlinks
-for file in $files; do
-  echo "Creating symlink to $file in home directory."
+for file in "${files[@]}"; do
+  if [ -L ~/$file ]; then
+    echo "Skip ~/$file, it's already a symbolic link"
+    continue
+  fi
+
+  if [ -e ~/$file ]; then 
+    echo "Moving ~/$file to ~/$file.bak"
+    mv ~/$file ~/$file.bak
+  fi
+
+  echo "Creating symlink to $dir/$file at ~/$file"
   ln -s $dir/$file ~/$file
-  echo "- - -"
 done
 
 # Create symlinks for VSCode
@@ -29,17 +34,7 @@ ln -s $dir/VSCode/settings.json ~/Library/Application\ Support/Code/User/setting
 ln -s $dir/VSCode/keybindings.json ~/Library/Application\ Support/Code/User/keybindings.json
 ln -s $dir/VSCode/snippets/ ~/Library/Application\ Support/Code/User/snippets
 
-# Create symlinks for zsh theme
-echo "Creating symlink for zsh configs."
-rm -rf ~/.oh-my-zsh/custom/themes
-ln -s ~/.dotfiles/oh-my-zsh/custom/themes ~/.oh-my-zsh/custom/themes
-ln -s ~/.dotfiles/oh-my-zsh/custom/zsh-autosuggestions.zsh ~/.oh-my-zsh/custom/zsh-autosuggestions.zsh
-
 # Lets make it look nicer
-echo ""
 echo "- - - - - - - - - -"
-echo ""
-echo "All done. Have a great day :-*"
-echo " "
+echo "All done. Have a great day."
 echo "- - - - - - - - - -"
-echo ""
